refactor: restructure file structure and remove unuse file.
